Move node diagnostics from backend to build-utils.
